refactor(home): optimize initial load
